Simplify the sequencer by passing generic values
The previous incarnation of the sequencing code was using closures to
move its sequence numbers in and out of the packet payload. This
introduced a lot of boilerplate code and increased complexity.

Replace it with a generics-based interface towards the sequencing and
put the payload in a serialized struct 'Payload'